+++ metadata-ingestion/src/datahub/utilities/urns/__init__.py
@@ -1,0 +1,2 @@
+from datahub.metadata._urns.urn_defs import *  # noqa: F401
+from datahub.utilities.urns._urn_base import Urn  # noqa: F401
