+++ metadata-ingestion/src/datahub/utilities/urns/corp_group_urn.py
@@ -1,1 +1,1 @@
-from datahub.utilities.urns.urn import CorpGroupUrn  # noqa: F401
+from datahub.utilities.urns import CorpGroupUrn  # noqa: F401
