+++ metadata-ingestion/src/datahub/utilities/urns/data_job_urn.py
@@ -1,1 +1,1 @@
-from datahub.utilities.urns.urn import DataJobUrn  # noqa: F401
+from datahub.utilities.urns import DataJobUrn  # noqa: F401
