+++ metadata-ingestion/src/datahub/utilities/urns/data_platform_urn.py
@@ -1,1 +1,1 @@
-from datahub.utilities.urns.urn import DataPlatformUrn  # noqa: F401
+from datahub.utilities.urns import DataPlatformUrn  # noqa: F401
