+++ metadata-ingestion/src/datahub/utilities/urns/data_process_instance_urn.py
@@ -1,1 +1,1 @@
-from datahub.utilities.urns.urn import DataProcessInstanceUrn  # noqa: F401
+from datahub.utilities.urns import DataProcessInstanceUrn  # noqa: F401
