+++ metadata-ingestion/src/datahub/utilities/urns/dataset_urn.py
@@ -1,3 +1,1 @@
-from datahub.utilities.urns.urn import DatasetUrn  # noqa: F401
-
-# TODO handling for platform_instance
+from datahub.utilities.urns import DatasetUrn  # noqa: F401
