+++ metadata-ingestion/src/datahub/utilities/urns/notebook_urn.py
@@ -1,1 +1,1 @@
-from datahub.utilities.urns.urn import NotebookUrn  # noqa: F401
+from datahub.utilities.urns import NotebookUrn  # noqa: F401
