+++ metadata-ingestion/src/datahub/utilities/urns/tag_urn.py
@@ -1,1 +1,1 @@
-from datahub.utilities.urns.urn import TagUrn  # noqa: F401
+from datahub.utilities.urns import TagUrn  # noqa: F401
